chore: organize rental list API and query files
