+++ src/api/index.ts
@@ -1,3 +1,2 @@
 export * from './refundAccount';
-export * from './api';
 export * from './rental';
